requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
python-jose[cryptography]
passlib[bcrypt]
certifi